+++ packages/electron-publish/package.json
@@ -24,7 +24,7 @@
     "lazy-val": "^1.0.5",
     "mime": "^2.5.2"
   },
-  "typings": "./out/publisher.d.ts",
+  "typings": "./out/index.d.ts",
   "devDependencies": {
     "@types/mime": "2.0.3"
   }
