+++ packages/electron-publish/src/gitHubPublisher.ts
@@ -5,9 +5,9 @@ import { ClientRequest } from "http"
 import { Lazy } from "lazy-val"
 import * as mime from "mime"
 import { parse as parseUrl, UrlWithStringQuery } from "url"
-import { getCiTag, PublishContext } from "./publisher"
+import { getCiTag } from "./publisher"
 import { HttpPublisher } from "./HttpPublisher"
-import { PublishOptions } from "./index"
+import { PublishContext, PublishOptions } from "./index"
 
 export interface Release {
   id: number
