+++ packages/electron-publish/src/index.ts
@@ -1,3 +1,7 @@
+import { CancellationToken } from "builder-util-runtime"
+import { MultiProgress } from "./multiProgress"
+import { Arch } from "builder-util"
+
 export type PublishPolicy = "onTag" | "onTagOrDraft" | "always" | "never"
 
 export { ProgressCallback } from "./progress"
@@ -6,5 +10,19 @@ export interface PublishOptions {
   publish?: PublishPolicy | null
 }
 
-export { PublishContext, UploadTask, Publisher, getCiTag } from "./publisher"
+export { Publisher, getCiTag } from "./publisher"
 export { HttpPublisher } from "./HttpPublisher"
+
+export interface PublishContext {
+  readonly cancellationToken: CancellationToken
+  readonly progress: MultiProgress | null
+}
+
+export interface UploadTask {
+  file: string
+  fileContent?: Buffer | null
+
+  arch: Arch | null
+  safeArtifactName?: string | null
+  timeout?: number | null
+}
